Change Refactor to Writer
