--- src/Refactor/Entity/WriteableEntityInteface.php
@@ -1,9 +1,0 @@
-<?php
-namespace DemandwareXml\Refactor\Entity;
-
-use DemandwareXml\Refactor\Xml\XmlWriter;
-
-interface WriteableEntityInteface
-{
-    public function write(XmlWriter $writer): void;
-}
