--- src/Refactor/Xml/XmlFormatterException.php
@@ -1,8 +1,0 @@
-<?php
-namespace DemandwareXml\Refactor\Xml;
-
-use Exception;
-
-class XmlFormatterException extends Exception
-{
-}
