+++ src/Writer/Entity/WriteableEntityInteface.php
@@ -1,0 +1,9 @@
+<?php
+namespace DemandwareXml\Writer\Entity;
+
+use DemandwareXml\Writer\Xml\XmlWriter;
+
+interface WriteableEntityInteface
+{
+    public function write(XmlWriter $writer): void;
+}
