+++ src/Writer/Xml/XmlFormatterException.php
@@ -1,0 +1,8 @@
+<?php
+namespace DemandwareXml\Writer\Xml;
+
+use Exception;
+
+class XmlFormatterException extends Exception
+{
+}
